# perf: add support for `AtomicMultiMarker` and `AtomicMarkerUnion` for `extra` markers