--- a/src/poetry/core/constraints/generic/constraint.py
+++ b/src/poetry/core/constraints/generic/constraint.py
@@ -137,7 +137,7 @@ class Constraint(BaseConstraint):
         return other.is_any()
 
     def invert(self) -> Constraint:
-        return Constraint(self._value, self._trans_op_inv[self.operator])
+        return self.__class__(self._value, self._trans_op_inv[self.operator])
 
     def difference(self, other: BaseConstraint) -> Constraint | EmptyConstraint:
         if other.allows(self):
@@ -207,8 +207,8 @@ class Constraint(BaseConstraint):
         return False
 
     def __eq__(self, other: object) -> bool:
-        if not isinstance(other, Constraint):
-            return NotImplemented
+        if not isinstance(other, self.__class__):
+            return False
 
         return (self.value, self.operator) == (other.value, other.operator)
 
@@ -220,3 +220,44 @@ class Constraint(BaseConstraint):
             return f"'{self._value}' {self._operator}"
         op = self._operator if self._operator != "==" else ""
         return f"{op}{self._value}"
+
+
+class ExtraConstraint(Constraint):
+    def __init__(self, value: str, operator: str = "==") -> None:
+        super().__init__(value, operator)
+        # Do the check after calling the super constructor,
+        # i.e. after the operator has been normalized.
+        if self._operator not in {"==", "!="}:
+            raise ValueError(
+                'Only the operators "==" and "!=" are supported for extra constraints'
+            )
+
+    def intersect(self, other: BaseConstraint) -> BaseConstraint:
+        from poetry.core.constraints.generic.multi_constraint import (
+            ExtraMultiConstraint,
+        )
+
+        if isinstance(other, Constraint):
+            if other == self:
+                return self
+
+            if self._value == other._value and self._operator != other.operator:
+                return EmptyConstraint()
+
+            return ExtraMultiConstraint(self, other)
+
+        return super().intersect(other)
+
+    def union(self, other: BaseConstraint) -> BaseConstraint:
+        from poetry.core.constraints.generic.union_constraint import UnionConstraint
+
+        if isinstance(other, Constraint):
+            if other == self:
+                return self
+
+            if self._value == other._value and self._operator != other.operator:
+                return AnyConstraint()
+
+            return UnionConstraint(self, other)
+
+        return super().union(other)