--- a/src/poetry/core/constraints/generic/parser.py
+++ b/src/poetry/core/constraints/generic/parser.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING
 
 from poetry.core.constraints.generic.any_constraint import AnyConstraint
 from poetry.core.constraints.generic.constraint import Constraint
+from poetry.core.constraints.generic.constraint import ExtraConstraint
 from poetry.core.constraints.generic.union_constraint import UnionConstraint
 from poetry.core.constraints.version.exceptions import ParseConstraintError
 
@@ -29,6 +30,17 @@ STR_CMP_CONSTRAINT = re.compile(
 
 @functools.cache
 def parse_constraint(constraints: str) -> BaseConstraint:
+    return _parse_constraint(constraints, Constraint)
+
+
+@functools.cache
+def parse_extra_constraint(constraints: str) -> BaseConstraint:
+    return _parse_constraint(constraints, ExtraConstraint)
+
+
+def _parse_constraint(
+    constraints: str, constraint_type: type[Constraint]
+) -> BaseConstraint:
     if constraints == "*":
         return AnyConstraint()
 
@@ -40,9 +52,13 @@ def parse_constraint(constraints: str) -> BaseConstraint:
 
         if len(and_constraints) > 1:
             for constraint in and_constraints:
-                constraint_objects.append(parse_single_constraint(constraint))
+                constraint_objects.append(
+                    _parse_single_constraint(constraint, constraint_type)
+                )
         else:
-            constraint_objects.append(parse_single_constraint(and_constraints[0]))
+            constraint_objects.append(
+                _parse_single_constraint(and_constraints[0], constraint_type)
+            )
 
         if len(constraint_objects) == 1:
             constraint = constraint_objects[0]
@@ -59,12 +75,14 @@ def parse_constraint(constraints: str) -> BaseConstraint:
         return UnionConstraint(*or_groups)
 
 
-def parse_single_constraint(constraint: str) -> Constraint:
+def _parse_single_constraint(
+    constraint: str, constraint_type: type[Constraint]
+) -> Constraint:
     # string comparator
     if m := STR_CMP_CONSTRAINT.match(constraint):
         op = m.group("op")
         value = m.group("value").strip()
-        return Constraint(value, op)
+        return constraint_type(value, op)
 
     # Basic comparator
 
@@ -75,6 +93,6 @@ def parse_single_constraint(constraint: str) -> Constraint:
 
         version = m.group(2).strip()
 
-        return Constraint(version, op)
+        return constraint_type(version, op)
 
     raise ParseConstraintError(f"Could not parse version constraint: {constraint}")